pyproject.toml
[project]
name = "mercado-envios-challenge"
version = "0.1.0"
description = "Prototype delivery promise optimization system"
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "lightgbm>=4.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0",
    "ipykernel>=6.25",
]
